+++ CHANGELOG.md
@@ -31,6 +31,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Custom images keys `case` and `str` in case config yaml file are renamed to `case_images` and `srt_variants_images`
 - Simplify and speed up case general report code
 - Speed up case retrieval in case_matching_causatives
+- Upgrade pymongo to version 4
 
 
 ## [4.71]